Consolidated logic so fetching and storing happens in the same module.
