+++ lib/screenplay/application.ex
@@ -6,20 +6,23 @@ defmodule Screenplay.Application do
   use Application
 
   def start(_type, _args) do
-    children = [
-      # Start the Telemetry supervisor
-      ScreenplayWeb.Telemetry,
-      # Start the PubSub system
-      {Phoenix.PubSub, name: Screenplay.PubSub},
-      # Start the Endpoint (http/https)
-      ScreenplayWeb.Endpoint,
-      # Start a worker by calling: Screenplay.Worker.start_link(arg)
-      # {Screenplay.Worker, arg}
-      Screenplay.OutfrontTakeoverTool.Alerts.State,
-      Screenplay.OutfrontTakeoverTool.Alerts.Reminders,
-      Screenplay.Scheduler,
-      Screenplay.Alerts.Cache.Supervisor
-    ]
+    children =
+      [
+        # Start the Telemetry supervisor
+        ScreenplayWeb.Telemetry,
+        # Start the PubSub system
+        {Phoenix.PubSub, name: Screenplay.PubSub},
+        # Start the Endpoint (http/https)
+        ScreenplayWeb.Endpoint,
+        Screenplay.OutfrontTakeoverTool.Alerts.State,
+        Screenplay.OutfrontTakeoverTool.Alerts.Reminders,
+        Screenplay.Scheduler
+      ] ++
+        if Application.get_env(:screenplay, :start_alerts_cache) do
+          [Screenplay.Alerts.Cache]
+        else
+          []
+        end
 
     # See https://hexdocs.pm/elixir/Supervisor.html
     # for other strategies and supported options
